remove bfv redundant module
